Significant refactor and additions
Started to merge in what I was doing with ctxswitch/vogon into the
Makefile.  Chef was a little unwieldy and heavy handed for managing my
day to day work on the laptop.  Refactored both naming and flow.  Added
contitionals to each target to perform specific tasks if run as a super
user or as an unpriv user.  This has grown into something a bit more
significant than just a dotfile repo, but keeping that name for now.